requirements.txt
subword-nmt
OpenNMT-tf
tensorflow-serving-api
